Provide dependencies to Mapbox using a basic 'service lookup via shared singleton' mini framework
